# refactor: Compress quotient table