Refactor how the quotient combination codeword is computed. Now it is done without ever constructing the entire quotient table, meaning that a lot less memory is used. Additionally, the part of the sum across constraint values of the same category now preceeds the multiplication by the zerofier-inverse, leading to less multiplications and better proving time.