@aszepieniec 2024-03-29
Benchmarking against master using `prove_halt` on my  laptop gives:
 - computing the quotient codeword goes from 12.35 ms to 8.91 ms (39% improvement)
 - overall proving time goes from 71.34 ms to 66.14 ms (7.8% improvement)
 - ram usage* goes from 76944 kB to 40472 kB (47% reduction)

*ram usage measured using `/bin/usr/time -v <cmd>` and looking at "Maximum resident set size".